% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcpActResources.R
\name{acpActResources}
\alias{acpActResources}
\title{Analisis de componentes principales para las diferentes variables}
\usage{
acpActResources(relation, variable = c("resources", "satisfaction",
  "influence", "aim"))
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}
}
\value{
Muestra un grafico de la variable (recursos, satisfaccion, influencia
         u objetivo) que se seleccione para el estudio y una tabla donde
         se puede observar la desviacion estandar, proporcion de la varianza
         explicada y la varianza acumulada para cada componente. Ademas
         origina una tabla donde se puede observar la desviacion estandar,
         proporcion de la varianza explicada por los componentes y la
         varianza acumulada.
}
\description{
Esta funcion muestra un grafico de la ubicacion de los recursos
             o actores con respecto a los dos ejes dominantes de componentes
             principales. La variable puede ser estado de los recursos
             (resources) o una de las variables caracteristicas de los
             actores: satisfaccion (satisfaction), influencia (influence) u
             objetivo (aim).
}
\examples{
data(potatoes)
acpActResources(potatoes, "satisfaction")
}
\author{
Maria Morales
}
\references{
Mardia, K. V., J. T. Kent and J. M. Bibby (1979). Multivariate
            Analysis, London: Academic Press.
}
\seealso{
pca, prcomp, ggplot
}

