% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisComponents.R
\name{analysisComponents}
\alias{analysisComponents}
\title{Analisis de componentes principales para los diferentes individuos}
\usage{
analysisComponents(relation, variable = c("resources", "satisfaction",
  "influence", "aim"))
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}
}
\value{
Muestra un grafico de la variable (recursos, satisfaccion, influencia
         u objetivo) que se seleccione para el estudio.
}
\description{
Esta funcion muestra un grafico de la ubicacion de los recursos
             o actores con respecto a los dos ejes dominantes de componentes
             principales. La variable puede ser estado de los recursos
             (resources) o una de las variables caracteristicas de los
             actores: satisfaccion (satisfaction), influencia (influence) u
             objetivo (aim).
}
\examples{
data(potatoes)
analysisComponents(potatoes, "satisfaction")
}
\author{
Maria Morales
}
\references{
Mardia, K. V., J. T. Kent and J. M. Bibby (1979). Multivariate
            Analysis, London: Academic Press.
}
\seealso{
pca, prcomp, ggplot
}

