% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxActResources.R
\name{boxActResources}
\alias{boxActResources}
\title{Graficos de caja para las variables caracteristicas de los actores o recursos}
\usage{
boxActResources(relation, variable = c("resources", "satisfaction",
  "influence", "aim"), perp = 2, tam = 0.6)
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}

\item{perp}{un valor entero que controla la orientacion en los ejes (0:
paralelo a los ejes, 1: horizontal, 2: perpendicular a los ejes,
3: vertical).}

\item{tam}{factor de amplificacion o disminucion del tamano del texto en los
ejes.}
}
\value{
Muestra un grafico de caja para la variable seleccionada (recursos,
        satisfaccion, influencia u objetivo). Los graficos de caja permiten
        resumir de manera visual estadisticos descriptivos basicos como
        tendencia central, dispersion y valores aberrantes o extremos.
}
\description{
Esta funcion realiza un grafico de caja o "box plot" para las
             variables que caracterizan a los actores (satisfaccion,
             influencia u objetivo) o el estado de los recursos.
}
\examples{
data(potatoes)
boxActResources(potatoes, "satisfaction")
}
\author{
Maria Morales
}
\references{
Chambers, J. M., Cleveland, W. S., Kleiner, B. and Tukey,
            P. A.(1983) Graphical Methods for Data Analysis.
            Wadsworth & Brooks/Cole
}
\seealso{
boxplot.stats, bxp
}

