% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxNbStep.R
\name{boxNbStep}
\alias{boxNbStep}
\title{Graficos de caja que comparan  numero de pasos, recursos y actores}
\usage{
boxNbStep(relation, variable = c("resources", "satisfaction", "influence",
  "aim"), perp = 2, tam = 0.6)
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}

\item{perp}{un valor entero que controla la orientacion en los ejes (0:
paralelo a los ejes, 1: horizontal, 2: perpendicular a los ejes,
3: vertical).}

\item{tam}{factor de amplificacion o disminucion del tamano del texto en los
ejes.}
}
\value{
Muestra tres graficos de caja: el primero con el numero de pasos en
        cada corrida, el segundo con los estados de los recursos y el tercero
        con la variable de interes de los actores (satisfaccion, influencia u
        objetivo). Los graficos de caja permiten resumir de manera visual
        estadisticos descriptivos basicos como tendencia central, dispersion
         y valores aberrantes o extremos.
}
\description{
Esta funcion realiza tres graficos de caja o "box plot"
             paralelos para comparar el numero de pasos en las corrida, el
             estado de los recursos y una de las variables que caracterizan a
             los actores (satisfaccion, influencia u objetivo).
}
\examples{
data(potatoes)
boxNbStep(potatoes, "satisfaction")
}
\author{
Maria Morales
}
\references{
Chambers, J. M., Cleveland, W. S., Kleiner, B. and Tukey,
            P. A.(1983) Graphical Methods for Data Analysis.
            Wadsworth & Brooks/Cole
}
\seealso{
boxplot.stats, bxp
}

