% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationActors.R
\name{correlationActors}
\alias{correlationActors}
\title{Analisis de Correlacion para las diferentes variables}
\usage{
correlationActors(relation, variable = c("resources", "satisfaction",
  "influence", "aim"), colors = (c("gray", "black")), cexl = 0.6)
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}

\item{colors}{define un vector de colores para la grafica.}

\item{cexl}{factor de amplificacion o disminucion del tamano del texto en
los ejes.}
}
\value{
Muestra un grafico para cada una de las variables (satisfaccion,
        influencia, objetivo y recursos) y una tabla donde se
        observa el indice de correlacion de las diferentes variables en
        estudio.
}
\description{
Esta funcion permite observar la correlacion que existe entre
             las diferentes variables que caracterizan a los actores
             (satisfaccion, influencia u objetivo) y el estado de los
             recursos, con respecto al numero de pasos.
}
\examples{
data(potatoes)
correlationActors(potatoes, "satisfaction")
}
\author{
Maria Morales
}
\references{
Murdoch, D.J. and Chow, E.D. (1996). A graphical display of large
            correlation matrices. The American Statistician 50, 178-180.
}
\seealso{
ellipse
}

