% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationActResources.R
\name{correlationActResources}
\alias{correlationActResources}
\title{Analisis de correlacion entre las variables caracteristicas de los actores y
el estado de los recursos}
\usage{
correlationActResources(relation, variable = c("satisfaction", "influence",
  "aim"), colors = (c("gray", "black")), cexl = 0.6)
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}

\item{colors}{define un vector de colores para la grafica.}

\item{cexl}{factor de amplificacion o disminucion del tamano del texto en los
ejes.}
}
\value{
Muestra un grafico para cada una de las variables (satisfaccion,
        influencia u objetivo) con respecto a la variable de estado de los
        recursos.
}
\description{
Esta funcion permite observar la correlacion que existe entre
             las variables que caracterizan a los actores (satisfaccion,
             influencia u objetivo) con respecto a la variable del estado de
             los recursos, tomando en cuenta el numero de pasos.
}
\examples{
data(potatoes)
correlationActResources(potatoes, "satisfaction")
}
\author{
Maria Morales
}
\references{
Murdoch, D.J. and Chow, E.D. (1996). A graphical display of large
            correlation matrices. The American Statistician 50, 178-180.
}
\seealso{
ellipse
}

