% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSocLab1.R
\name{dataSocLab1}
\alias{dataSocLab1}
\title{Lectura del archivo de salida de SocLab para el analisis de sensibilidad}
\usage{
dataSocLab1(direction_file)
}
\arguments{
\item{direction_file}{es una cadena de caracteres que contiene el nombre o el
camino (directorio + nombre) del archivo de entrada, donde esta el
archivo con los datos de salida de SocLab.}
}
\value{
Retorna una lista  que contiene:

 ResName: vector que contiene los nombres de los recursos utilizados en el
          modelo.

 ActName: vector que contiene los nombres de los actores utilizados en el
          modelo.

 staterel: marco de datos con el estado de los recursos para todas las
            corridas del modelo.

 satis: marco de datos con la variable satisfaccion de cada uno de los
               actores para todas las corridas del modelo.

 inf: marco de datos con la variable influencia de cada uno de los
            actores para todas las corridas del modelo.

 aim: marco de datos con la variable objetivo de cada uno de los actores para
      todas las corridas del modelo.

 param0: vector con la variacion de los parametros en cada corrida.

 nb_step: vector con el numero de pasos en cada corrida.

 num_exp: vector con el numero de experimentos en cada corrida.

 parameters: parametro a variar.
}
\description{
Esta funcion es para leer los datos de salida de SocLab. Se presume que los
     datos estan en un formato texto separado por tabulaciones. Si los datos
     no se encuentran en el directorio de trabajo hay que especificar el
     camino completo.
}
\examples{
 ## Un ejemplo del archivo de salida de SocLab es
 ## sensitivityPotatoesModelV4.txt que se
 ## encuentra en el directorio extdata del directorio de instalacion del
 ## paquete AESS
 ## La siguiente instruccion devuelve como una cadena de caracteres la
 ## direccion (camino + nombre) de este archivo
 nombre.arc=system.file("extdata", "sensitivityPotatoesModelV4.txt",
 package = "AESS")
 ## La entrada a la funcion es la direccion donde se encuentra el archivo
 datos<-dataSocLab1(nombre.arc)
}

