% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionActors.R
\name{regressionActors}
\alias{regressionActors}
\title{Analisis de regresion lineal multiple para las variables caracteristicas de
los actores}
\usage{
regressionActors(relation, varDep, variable = c("satisfaction", "influence",
  "aim"))
}
\arguments{
\item{relation}{la lista con la salidas del modelo.}

\item{varDep}{la variable dependiente.}

\item{variable}{la variable a analizar.}
}
\value{
Muestra un conjunto de graficas donde se observa la relacion
        entre una variable denominada Y con respecto al vector de variables
        explicativas  (satisfaccion, influencia u objetivo). Ademas, las
        graficas de diagnostico y una tabla  donde se puede observar algunos
        estadisticos como el coeficiente de correlacion de Pearson, el
        coeficiente de determinacion R2 ajustado entre otros.
}
\description{
Esta funcion permite modelar la relacion entre una de las
             variables que caracterizan a los actores (satisfaccion,
             influencia u objetivo) denominada Y y un vector de variables
             explicativas (satisfaccion, influencia u objetivo) denominadas X
}
\examples{
data(potatoes)
regressionActors(potatoes, "PROINPA_MST", "satisfaction")
}
\author{
Maria Morales
}
\references{
Chambers, J. M. (1992) Linear models. Chapter 4 of Statistical
            Models in S eds J. M. Chambers and T. J. Hastie,
            Wadsworth & Brooks/Cole.

            Wilkinson, G. N. and Rogers, C. E. (1973) Symbolic descriptions
            of factorial models for analysis of variance. Applied Statistics.
}
\seealso{
lm, summary
}

