% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensibilityAnalysis.R
\name{sensibilityAnalysis}
\alias{sensibilityAnalysis}
\title{Analisis de sensibilidad para las diferentes variables}
\usage{
sensibilityAnalysis(lp, variable = c("satis", "inf", "aim", "staterel"),
  opcion = c("param0", "nb_step", "num_exp"), j = 0.5, k = 35)
}
\arguments{
\item{lp}{la lista con la salidas del modelo.}

\item{variable}{la variable a analizar.}

\item{opcion}{una variable que almacena una cadena de caracteres. Esta cadena
de caracteres contiene: el parametro (param0), el numero de
pasos (nb_step) y el numero de experimentos (num_exp) respectivamente.}

\item{j}{un valor que define la posicion de la leyenda en el eje x.}

\item{k}{un valor que define la posicion de la leyenda en el eje y.}
}
\value{
Muestra un grafico de la variable que caracterizan a los actores
        (satisfaccion, influencia u objetivo) y la variable recursos que
        controla cada actor, asi como tambien un grafico para el nb_step y
        num_exp.
}
\description{
Esta funcion permite determinar que tan sensible puede llegar a
             ser el modelo ante cambios realizados en los valores de los
             parametros.
}
\examples{
data(parameters)
sensibilityAnalysis(parameters, "satis", "param0")
}
\author{
Maria Morales
}

